Split [pypi] tests
